+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Fixed
+
+- Accept an initial capital letter when both the question and the answer of a quiz are in lower case. Fixes [#671](https://github.com/fniessink/toisto/issues/671).
+
 ### Added
 
 - Add support for the present perfect tense. Closes [#632](https://github.com/fniessink/toisto/issues/632).
 - When showing examples, also show the meaning of the examples. Closes [#638](https://github.com/fniessink/toisto/issues/638).
+- Add several concepts.
 
 ## 0.18.1 - 2024-04-09